Feature update: Issue #16
Extract Search toolbar into a composable component
